requirements.txt
# ── CLI bağımlılıkları ─────────────────────────────────────────────
click>=8.1.0
rich>=13.0.0
colorama>=0.4.6
pyyaml>=6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
robotframework>=6.0

# ── Web UI bağımlılıkları ──────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# ── Slack entegrasyonu ─────────────────────────────────────────────
requests>=2.31.0
